+++ SCT2/trunk/plugins/org.yakindu.sct.model.stext/src-gen/org/yakindu/sct/model/stext/stext/impl/TimeEventSpecImpl.java
@@ -2,6 +2,7 @@
  * <copyright>
  * </copyright>
  *
+
  */
 package org.yakindu.sct.model.stext.stext.impl;
 
@@ -61,7 +62,7 @@ public class TimeEventSpecImpl extends EventSpecImpl implements TimeEventSpec
    * @generated
    * @ordered
    */
-  protected static final int VALUE_EDEFAULT = 0;
+  protected static final long VALUE_EDEFAULT = 0L;
 
   /**
    * The cached value of the '{@link #getValue() <em>Value</em>}' attribute.
@@ -71,7 +72,7 @@ public class TimeEventSpecImpl extends EventSpecImpl implements TimeEventSpec
    * @generated
    * @ordered
    */
-  protected int value = VALUE_EDEFAULT;
+  protected long value = VALUE_EDEFAULT;
 
   /**
    * The default value of the '{@link #getUnit() <em>Unit</em>}' attribute.
@@ -142,7 +143,7 @@ public class TimeEventSpecImpl extends EventSpecImpl implements TimeEventSpec
    * <!-- end-user-doc -->
    * @generated
    */
-  public int getValue()
+  public long getValue()
   {
     return value;
   }
@@ -152,9 +153,9 @@ public class TimeEventSpecImpl extends EventSpecImpl implements TimeEventSpec
    * <!-- end-user-doc -->
    * @generated
    */
-  public void setValue(int newValue)
+  public void setValue(long newValue)
   {
-    int oldValue = value;
+    long oldValue = value;
     value = newValue;
     if (eNotificationRequired())
       eNotify(new ENotificationImpl(this, Notification.SET, StextPackage.TIME_EVENT_SPEC__VALUE, oldValue, value));
@@ -217,7 +218,7 @@ public class TimeEventSpecImpl extends EventSpecImpl implements TimeEventSpec
         setType((TimeEventType)newValue);
         return;
       case StextPackage.TIME_EVENT_SPEC__VALUE:
-        setValue((Integer)newValue);
+        setValue((Long)newValue);
         return;
       case StextPackage.TIME_EVENT_SPEC__UNIT:
         setUnit((TimeUnit)newValue);
